Moved static methods and dependencies out of Piece.java into PieceFactory in new package game; moved Queue into package game
