+++ game/Queue.java
@@ -1,4 +1,6 @@
-package hex;
+package game;
+
+import hex.Piece;
 
 /**
  * Represents a simple fixed-length queue of {@code Piece} objects.
@@ -17,7 +19,7 @@ package hex;
  * externally-defined elements. New pieces are created via an
  * internal {@link #generate()} method.
  * @author William Wu
- * @version 1.1
+ * @version 1.2
  */
 public class Queue{
     /** Internal fixed-size array of Piece elements. */
@@ -119,21 +121,21 @@ public class Queue{
     /**
      * Generates a new {@link Piece} according to piece generation logic.
      * @return a newly generated {@code Piece}.
-     * @see Piece#generatePiece()
+     * @see PieceFactory#generatePiece()
      * @see #generate(Piece)
      */
     protected Piece generate(){
-        return (Piece) pieceProcessor.process(new Object[]{Piece.generatePiece(), Launcher.LaunchEssentials.isEasyMode(), GUI.GameEssentials.engine()})[0];
+        return (Piece) pieceProcessor.process(new Object[]{PieceFactory.generatePiece(), Launcher.LaunchEssentials.isEasyMode(), GUI.GameEssentials.engine()})[0];
     }
     /**
      * Generates a new {@link Piece} according to piece generation logic.
      * @param dequeued the {@code Piece} that is dequeued from the queue.
      * @return a newly generated {@code Piece}.
-     * @see Piece#generatePiece()
+     * @see PieceFactory#generatePiece()
      * @see #generate()
      */
     protected Piece generate(Piece dequeued){
-        return (Piece) pieceProcessor.process(new Object[]{Piece.generatePiece(), Launcher.LaunchEssentials.isEasyMode(), GUI.GameEssentials.engine(), dequeued})[0];
+        return (Piece) pieceProcessor.process(new Object[]{PieceFactory.generatePiece(), Launcher.LaunchEssentials.isEasyMode(), GUI.GameEssentials.engine(), dequeued})[0];
     }
     /**
      * Returns the first {@link Piece} in the queue without removing it.
